Add site layout has been updated to the point where i'm almost happy for it to be completed. Popup layout over map is completed.
Add Site now includes the function to add multiple images and they now save to the database, the sizing of the images on the add site screen however need to be adjusted as they are far too large.

Job for next time: Create Fetch Images class that collects the images for a specific site which can be run as site opens.